Refactor RightClickMenu into sub-components for better maintainability
- Extracted `RightClickMenu` from a single large file into a directory structure.
- Created `StandardSettings`, `CarouselSettings`, `CoverFlowSettings` for view-specific logic.
- Extracted reusable `TopActionRow`, `ViewModeSelector`, and `ButtonColorEditor` components.
- Centralized constants and types.
- Preserved existing functionality and logic.
