+++ renderer/src/components/RightClickMenu/constants.ts
@@ -1,0 +1,38 @@
+export const sliderDefaults = {
+  gridSize: 120,
+  logoSize: 100,
+  listSize: 128,
+  detailsBarSize: 14,
+  selectedBoxArtSize: 25,
+  gameTilePadding: 3,
+  backgroundBlur: 40,
+  backgroundBrightnessPercent: 30,
+  carouselLogoSize: 100,
+  carouselDescriptionSize: 18,
+  carouselButtonSize: 14,
+  categoriesTopSize: 12,
+  logoBackgroundOpacity: 100,
+  listTileHeight: 128,
+  listBoxartSize: 96,
+  listLogoSize: 96,
+  listTitleTextSize: 18,
+  listSectionTextSize: 14,
+  panelWidth: 800,
+  fanartHeight: 320,
+  descriptionWidth: 50,
+  perGameLogoSize: 100,
+  rightPanelBoxartSize: 120,
+  rightPanelTextSize: 14,
+  rightPanelButtonSize: 14,
+  detailsPanelOpacity: 80,
+  coverFlowCoverSize: 300,
+  coverFlowReflection: 60,
+  coverFlowVerticalOffset: 0,
+  coverFlowSideOpacity: 100,
+};
+
+export const defaultButtonColors = {
+  playColor: '#0ea5e9',
+  editColor: '#6b7280',
+  modManagerColor: '#a855f7',
+};
