+++ renderer/src/components/RightClickMenu/types.ts
@@ -1,0 +1,150 @@
+import type { Game } from '../../types/game';
+
+export interface RightClickMenuProps {
+  x: number;
+  y: number;
+  onClose: () => void;
+  viewMode: 'grid' | 'list' | 'logo' | 'carousel' | 'coverflow';
+  onViewModeChange?: (mode: 'grid' | 'list' | 'logo' | 'carousel' | 'coverflow') => void;
+  activeGame?: Game;
+  onActiveGameChange?: (game: Game) => void;
+  gridSize?: number;
+  onGridSizeChange?: (size: number) => void;
+  logoSize?: number;
+  onLogoSizeChange?: (size: number) => void;
+  listSize?: number;
+  onListSizeChange?: (size: number) => void;
+  gameTilePadding?: number;
+  onGameTilePaddingChange?: (padding: number) => void;
+  backgroundBlur?: number;
+  onBackgroundBlurChange?: (blur: number) => void;
+  backgroundBrightness?: number;
+  onBackgroundBrightnessChange?: (brightness: number) => void;
+  selectedBoxArtSize?: number;
+  onSelectedBoxArtSizeChange?: (size: number) => void;
+  carouselLogoSize?: number;
+  onCarouselLogoSizeChange?: (size: number) => void;
+  detailsBarSize?: number;
+  onDetailsBarSizeChange?: (size: number) => void;
+  showCarouselDetails?: boolean;
+  onShowCarouselDetailsChange?: (show: boolean) => void;
+  showCarouselLogos?: boolean;
+  onShowCarouselLogosChange?: (show: boolean) => void;
+  carouselButtonSize?: number;
+  onCarouselButtonSizeChange?: (size: number) => void;
+  carouselDescriptionSize?: number;
+  onCarouselDescriptionSizeChange?: (size: number) => void;
+  carouselDescriptionAlignment?: 'left' | 'center' | 'right';
+  onCarouselDescriptionAlignmentChange?: (alignment: 'left' | 'center' | 'right') => void;
+  carouselButtonAlignment?: 'left' | 'center' | 'right';
+  onCarouselButtonAlignmentChange?: (alignment: 'left' | 'center' | 'right') => void;
+  carouselLogoAlignment?: 'left' | 'center' | 'right';
+  onCarouselLogoAlignmentChange?: (alignment: 'left' | 'center' | 'right') => void;
+  listViewOptions?: {
+    showDescription: boolean;
+    showCategories: boolean;
+    showPlaytime: boolean;
+    showReleaseDate: boolean;
+    showGenres: boolean;
+    showPlatform: boolean;
+    showLauncher?: boolean;
+    showLogos?: boolean;
+    titleTextSize?: number;
+    displayMode?: 'boxart-title' | 'logo-title' | 'logo-only' | 'title-only' | 'icon-title';
+    sectionTextSize?: number;
+    tileHeight?: number;
+    boxartSize?: number;
+    logoSize?: number;
+  };
+  onListViewOptionsChange?: (options: {
+    showDescription: boolean;
+    showCategories: boolean;
+    showPlaytime: boolean;
+    showReleaseDate: boolean;
+    showGenres: boolean;
+    showPlatform: boolean;
+    showLauncher?: boolean;
+    showLogos?: boolean;
+    titleTextSize?: number;
+    displayMode?: 'boxart-title' | 'logo-title' | 'logo-only' | 'title-only' | 'icon-title';
+    sectionTextSize?: number;
+    tileHeight?: number;
+    boxartSize?: number;
+    logoSize?: number;
+  }) => void;
+  // Grid view specific props
+  showLogoOverBoxart?: boolean;
+  onShowLogoOverBoxartChange?: (show: boolean) => void;
+  logoPosition?: 'top' | 'middle' | 'bottom' | 'underneath';
+  onLogoPositionChange?: (position: 'top' | 'middle' | 'bottom' | 'underneath') => void;
+  logoBackgroundColor?: string;
+  onLogoBackgroundColorChange?: (color: string) => void;
+  logoBackgroundOpacity?: number;
+  onLogoBackgroundOpacityChange?: (opacity: number) => void;
+  gridDescriptionSize?: number;
+  onGridDescriptionSizeChange?: (size: number) => void;
+  gridButtonSize?: number;
+  onGridButtonSizeChange?: (size: number) => void;
+  gridButtonLocation?: 'left' | 'middle' | 'right';
+  onGridButtonLocationChange?: (location: 'left' | 'middle' | 'right') => void;
+  // Right panel (GameDetailsPanel) specific props
+  rightPanelLogoSize?: number;
+  onRightPanelLogoSizeChange?: (size: number) => void;
+  rightPanelBoxartPosition?: 'left' | 'right' | 'none';
+  onRightPanelBoxartPositionChange?: (position: 'left' | 'right' | 'none') => void;
+  rightPanelBoxartSize?: number;
+  onRightPanelBoxartSizeChange?: (size: number) => void;
+  rightPanelTextSize?: number;
+  onRightPanelTextSizeChange?: (size: number) => void;
+  rightPanelButtonSize?: number;
+  onRightPanelButtonSizeChange?: (size: number) => void;
+  rightPanelButtonLocation?: 'left' | 'middle' | 'right';
+  onRightPanelButtonLocationChange?: (location: 'left' | 'middle' | 'right') => void;
+  detailsPanelOpacity?: number;
+  onDetailsPanelOpacityChange?: (opacity: number) => void;
+  // Panel width for saving/restoring divider position
+  panelWidth?: number;
+  onPanelWidthChange?: (width: number) => void;
+  // Game details panel divider height control
+  fanartHeight?: number;
+  onFanartHeightChange?: (height: number) => void;
+  descriptionWidth?: number;
+  onDescriptionWidthChange?: (width: number) => void;
+  showCategoriesInGameList?: boolean;
+  onShowCategoriesInGameListChange?: (show: boolean) => void;
+  categoriesPosition?: 'top' | 'bottom';
+  onCategoriesPositionChange?: (position: 'top' | 'bottom') => void;
+  categoriesTopAlignment?: 'left' | 'center' | 'right';
+  onCategoriesTopAlignmentChange?: (alignment: 'left' | 'center' | 'right') => void;
+  categoriesTopSize?: number;
+  onCategoriesTopSizeChange?: (size: number) => void;
+  isViewFlipped?: boolean;
+  onViewFlipChange?: (flipped: boolean) => void;
+  // Button colors per view
+  rightPanelButtonColors?: { playColor?: string; editColor?: string; modManagerColor?: string };
+  onRightPanelButtonColorsChange?: (colors: { playColor?: string; editColor?: string; modManagerColor?: string }) => void;
+  carouselButtonColors?: { playColor?: string; editColor?: string; modManagerColor?: string };
+  onCarouselButtonColorsChange?: (colors: { playColor?: string; editColor?: string; modManagerColor?: string }) => void;
+  gridButtonColors?: { playColor?: string; editColor?: string; modManagerColor?: string };
+  onGridButtonColorsChange?: (colors: { playColor?: string; editColor?: string; modManagerColor?: string }) => void;
+  listButtonColors?: { playColor?: string; editColor?: string; modManagerColor?: string };
+  onListButtonColorsChange?: (colors: { playColor?: string; editColor?: string; modManagerColor?: string }) => void;
+  logoButtonColors?: { playColor?: string; editColor?: string; modManagerColor?: string };
+  onLogoButtonColorsChange?: (colors: { playColor?: string; editColor?: string; modManagerColor?: string }) => void;
+  // Cover Flow only (simplified menu)
+  coverFlowCoverSize?: number;
+  onCoverFlowCoverSizeChange?: (size: number) => void;
+  coverFlowReflection?: number;
+  onCoverFlowReflectionChange?: (value: number) => void;
+  coverFlowVerticalOffset?: number;
+  onCoverFlowVerticalOffsetChange?: (value: number) => void;
+  coverFlowSideOpacity?: number;
+  onCoverFlowSideOpacityChange?: (value: number) => void;
+  coverFlowShowButtons?: boolean;
+  onCoverFlowShowButtonsChange?: (show: boolean) => void;
+  coverFlowButtonPosition?: 'left' | 'middle' | 'right';
+  onCoverFlowButtonPositionChange?: (pos: 'left' | 'middle' | 'right') => void;
+  coverFlowButtonColors?: { playColor?: string; editColor?: string; modManagerColor?: string };
+  onCoverFlowButtonColorsChange?: (colors: { playColor?: string; editColor?: string; modManagerColor?: string }) => void;
+  onSettingsImported?: () => void;
+}
